STD-40 | Realize TheoryBlock Controller (#37)
* Query Task added

* created: // AddTheoryBlockCommand.cs
// GetTheoryBlockByIdQuery.cs
// RemoveTheoryBlockByIdCommand.cs

* Added course controller and course requests and responses

* Fixed http get/post

* Added course mapper profile

* Fixwd conversions
#bodyazaebal

* Argument Fix

* AddTheoryBlockRequest, RemoveTheoryBlockByIdRequest and TheoryBlockController created

* Fix Equals and code style

* HttpGet by Id added into TheoryBlockController.cs and other minor changes
Added:
GetTheoryBlockByIdRequest
GetTheoryBlockByIdResponse
GetTheoryBlocksRequest
GetTheoryBlocksResponse
TheoryBlockProfile

* After review update

* Remove stupid generics

* GetTheoryBlockByCourseId pre-update
Add GetTheoryBlockByCourseIdRequest and GetTheoryBlockByCourseIdResponse
Add GetTheoryBlockByCourseId logic to the TheoryBlockController.cs, now commented due to lack of some queries

* Add "Update" functionality to the TheoryBlockController

* Add GetTheoryBlockByCourseId
Add GetTheoryBlockByCourseIdQuery
Uncomment code in TheoryBlockController.cs

* Pre-rewrite update

* Major update of the TheoryBlockController
TheoryBlockController now works like a part of Course
Remove GetTheoryBlocksRequest.cs due to is uselessness
Other minor changes due to adding CourseId validation

* GetTheoryBlockByCourseIdQuery returns a List of TheoryBlocks if theoryBlock == null

* Pull update from dev

* Add Course validation
In AddTheoryBlockCommand, GetTheoryBlockByCourseIdQuery, GetTheoryBlockByIdQuery, RemoveTheoryBlockByIdCommand, UpdateTheoryBlockCommand

* Remove useless comment

* Minor changes
Delete GetTheoryBlocksResponse.cs
GetTheoryBlockByCourseId => GetTheoryBlocksByCourseId

* Remove unnecessary Course elements

* Tests 1/2
Add row in mapper profile
Rename and minor changes in GetTheoryBlocksByCourseIdQuery.cs

* Tests 2/2
Some changes in Update and Remove blocks

* Tests 3/2
Fixed CourseId validation in UpdateTheoryBlockCommand.cs
Update now works correctly
Other minor changes

* Add now works, rewrite operation result text in GetTheoryBlockByIdQuery
Other minor changes

* Rework courseId exception in Remove, GetById, GetByCourseId, rework exception texts in GetByCourseId

* Unnecessary comments was removed

* Change course to courseExistent

+++ server/StudySharp.ApplicationServices/Commands/RemoveTheoryBlockByIdCommand.cs
@@ -1,6 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 using StudySharp.Domain.Constants;
 using StudySharp.Domain.General;
 using StudySharp.Domain.Models;
@@ -11,6 +12,7 @@ namespace StudySharp.ApplicationServices.Commands
     public sealed class RemoveTheoryBlockByIdCommand : IRequest<OperationResult>
     {
         public int Id { get; set; }
+        public int CourseId { get; set; }
     }
 
     public sealed class RemoveTheoryBlockByIdCommandHandler : IRequestHandler<RemoveTheoryBlockByIdCommand, OperationResult>
@@ -22,15 +24,19 @@ namespace StudySharp.ApplicationServices.Commands
             _context = sharpDbContext;
         }
 
-        public async Task<OperationResult> Handle(
-            RemoveTheoryBlockByIdCommand request,
-            CancellationToken cancellationToken)
+        public async Task<OperationResult> Handle(RemoveTheoryBlockByIdCommand request, CancellationToken cancellationToken)
         {
-            var theoryBlock = await _context.TheoryBlocks.FindAsync(request.Id, cancellationToken);
+            var courseExistent = await _context.Courses.AnyAsync(_ => _.Id == request.CourseId, cancellationToken);
+            if (!courseExistent)
+            {
+                return OperationResult.Fail(string.Format(ErrorConstants.EntityNotFound, nameof(Course), nameof(Course.Id), request.CourseId));
+            }
+
+            var theoryBlock = await _context.TheoryBlocks.FirstOrDefaultAsync(_ => _.Id == request.Id, cancellationToken);
 
             if (theoryBlock == null)
             {
-                return OperationResult.Fail(string.Format(ErrorConstants.EntityNotFound, nameof(TheoryBlock), nameof(TheoryBlock.Name), request.Id));
+                return OperationResult.Fail(string.Format(ErrorConstants.EntityNotFound, nameof(TheoryBlock), nameof(TheoryBlock.Id), request.Id));
             }
 
             _context.TheoryBlocks.Remove(theoryBlock);
